+++ src/index.ts
@@ -1,25 +1,21 @@
 import { readFileSync } from "fs";
-import { transform } from "esbuild";
 import { Plugin } from "vite";
 
-export function svgPlugin(opts?: { useInnerHTML?: boolean }): Plugin {
+export function svgPlugin(): Plugin {
+  let production = false;
   return {
     name: "svg",
     enforce: "pre",
+    config(_, env) {
+      production = env.command === "build";
+    },
     async load(id) {
       if (id.endsWith(".svg")) {
-        const { code, warnings } = await transform(
-          svgToJSX(readFileSync(id, "utf-8"), opts?.useInnerHTML),
-          { loader: "jsx" }
-        );
-        for (const warning of warnings) {
-          console.log(warning.location, warning.text);
-        }
-        return code;
+        return svgToJS(readFileSync(id, "utf-8"), production);
       }
       if (id.endsWith(".svg?inline")) {
         const base64 = Buffer.from(
-          readFileSync(id.replace("?inline", ""), "utf-8")
+          readFileSync(id.replace("?inline", ""), "utf-8"),
         ).toString("base64");
         return `export default "data:image/svg+xml;base64,${base64}"`;
       }
@@ -27,30 +23,36 @@ export function svgPlugin(opts?: { useInnerHTML?: boolean }): Plugin {
   };
 }
 
-export const svgToJSX = (svg: string, useInnerHTML?: boolean) => {
-  let jsx: string;
-  if (useInnerHTML) {
-    const index = svg.indexOf(">");
-    const content = svg
-      .slice(index + 1, svg.indexOf("</svg>"))
-      .trim()
-      .replace(/\s+/g, " ");
-    jsx = `${updatePropsCase(
-      svg.slice(0, index)
-    )} ref={ref} {...props} dangerouslySetInnerHTML={{ __html: '${content}' }} />`;
-  } else {
-    jsx = updatePropsCase(svg).replace(">", " ref={ref} {...props}>");
+const attributesRE = /\s([a-zA-Z0-9-:]+)=("[^"]*")/gu;
+
+export const svgToJS = (svg: string, production: boolean) => {
+  const index = svg.indexOf(">");
+  const content = svg
+    .slice(index + 1, svg.indexOf("</svg>"))
+    .trim()
+    .replace(/\s+/g, " ");
+  let attributes = "";
+  for (const match of svg.slice(0, index).matchAll(attributesRE)) {
+    attributes += `    ${transformKey(match[1])}: ${match[2]},\n`;
   }
-  return `import React from "react";const ReactComponent = React.forwardRef((props, ref) => (${jsx}));export default ReactComponent;`;
+  const jsxImport = production
+    ? 'import { jsx } from "react/jsx-runtime";'
+    : 'import { jsxDEV as jsx } from "react/jsx-dev-runtime";';
+  return `${jsxImport}
+import { forwardRef } from "react";
+export default forwardRef((props, ref) => jsx("svg", {
+${attributes}    ref,
+    ...props,
+    dangerouslySetInnerHTML: { __html: '${content}' }
+  })
+);`;
 };
 
-const updatePropsCase = (svg: string) =>
-  svg.replace(/\s([a-z-:]*)="[^"]*"/gu, (string, key: string) => {
-    if (key.startsWith("data-")) return string;
-    const keyWithoutDashes = camelCaseOn(key, "-");
-    const keyWithoutDots = camelCaseOn(keyWithoutDashes, ":");
-    return string.replace(key, keyWithoutDots);
-  });
+const transformKey = (key: string) => {
+  if (key.startsWith("data-")) return `"${key}"`;
+  const keyWithoutDashes = camelCaseOn(key, "-");
+  return camelCaseOn(keyWithoutDashes, ":");
+};
 
 const camelCaseOn = (string: string, delimiter: string) =>
   string